reorganized and fix backend bugs
